+++ frontend/src/pages/WritingArticles/index.tsx
@@ -1,62 +1,27 @@
-import { AxiosResponse, AxiosError } from 'axios';
-import { useEffect, useRef, useState } from 'react';
-import { useMutation } from 'react-query';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
-import { postWritingArticle } from '@/api/article';
 import Loading from '@/components/common/Loading/Loading';
 import PageLayout from '@/components/layout/PageLayout/PageLayout';
-import { CATEGORY } from '@/constants/categoryType';
 import ToastUiEditor from '@/pages/WritingArticles/ToastUiEditor/ToastUiEditor';
+
+import usePostWritingArticles from '@/pages/WritingArticles/hooks/usePostWritingArticles';
+
 import * as S from '@/pages/WritingArticles/index.styles';
-import { Editor } from '@toast-ui/react-editor';
 
 const WritingArticles = () => {
 	const { category } = useParams();
 
-	const { data, mutate, isError, isLoading, isSuccess, error } = useMutation<
-		AxiosResponse<{ id: string }>,
-		AxiosError,
-		{ title: string; category: string; content: string }
-	>(postWritingArticle);
-
-	const [title, setTitle] = useState('');
-	const [categoryOption, setCategoryOption] = useState(category ? category : '');
-	const content = useRef<Editor | null>(null);
-	const navigate = useNavigate();
-
-	useEffect(() => {
-		if (isSuccess && category === CATEGORY.discussion) {
-			if (confirm('글 등록이 완료되었습니다. 투표를 등록하시겠습니까?')) {
-				navigate(`/votes/${data.data.id}`);
-				return;
-			}
-			navigate(`/articles/${category}/${data.data.id}`);
-		}
-		if (isSuccess && category === CATEGORY.question) {
-			navigate(`/articles/${category}/${data.data.id}`);
-		}
-	}, [isSuccess]);
-
-	const handleSubmitButtonClick = () => {
-		if (content.current === null) {
-			return;
-		}
-		mutate({
-			title,
-			category: categoryOption,
-			content: content.current.getInstance().getMarkdown(),
-		});
-	};
+	const {
+		isLoading,
+		content,
+		handleSubmitButtonClick,
+		title,
+		setTitle,
+		categoryOption,
+		setCategoryOption,
+	} = usePostWritingArticles(category);
 
 	if (isLoading) return <Loading />;
-
-	if (isError) {
-		if (error instanceof Error) {
-			return <div>{error.message}</div>;
-		}
-		return null;
-	}
 
 	return (
 		<S.Container>
@@ -96,7 +61,10 @@ const WritingArticles = () => {
 			<S.Content>
 				<ToastUiEditor initContent={''} ref={content} />
 			</S.Content>
-			<S.SubmitButton type="button" onClick={handleSubmitButtonClick}>
+			<S.SubmitButton
+				type="button"
+				onClick={() => handleSubmitButtonClick({ title, categoryOption })}
+			>
 				등록하기
 			</S.SubmitButton>
 		</S.Container>
